fix: preserve non-maximized config for autohide panels/docks
This PR fixes the behavior of autohide panels/docks when a maximized window is focused.

Previously, the panel/dock would expand itself to the edges, regardless of whether the setting is turned off in the settings. While this is appropriate for regular panels, autohide panels/docks should not expand to edges unless it's set that way in settings.