Initial support for passing touch input to clients
Updated from https://github.com/pop-os/cosmic-panel/pull/199 and
https://github.com/pop-os/xdg-shell-wrapper/pull/22.

Probably better to merge some version of this and improve latter,
instead of leaving it unmerged indefinitely.

But it seems to crash sometimes trying to change workspaces with this?
Exiting with status code 101? I saw that with another cosmic-panel
change, but not sure what's going on there (that's different from
SIGABRT or other signals killing the process...)