Use `space_focus()` for overflow popups as well
This changes a few subtle things, but as far as I can tell should be
correct.

This adds a `PanelSpaceElement` trait, and also a
`TryFrom<PopupMappedInternal> for SpaceTarget` implementation, so that
`space_focus()` can be generic and accept both `PopupSpaceElement` and
`PanelSpaceElement` spaces.

Previously, popup elements used `bbox()` instead of `geometry()`, and
didn't adjust that for scale.  It also didn't clamp to the configured
size. I assume this should instead match the implementation used for
panel elements.

Meanwhile, only the popup element bbox check considered `.geometry().loc`.
Presumably both should.