Define `visiblity` in `PanelSpace::new()` instead of accepting as arg
This argument was set the same way, except in the two uses in
`update_space()`. But currently, that function appears to be called only
on config changes? (Given `PanelCalloopMsg::RestartSpace` isn't used
anywhere). Which presumably shouldn't be behaving differently from uses
elsewhere?

Also one of those uses is specific to `CosmicPanelOutput::Active`, and
that feature doesn't seem to be working currently, as far as I can tell
(but also isn't shown in cosmic-settings, so it's not urgent to fix).